FlowFeatureMapper changes removed
