+++ src/main/java/org/broadinstitute/hellbender/tools/walkers/featuremapping/SNVMapper.java
@@ -23,34 +23,26 @@ public class SNVMapper implements FeatureMapper {
 
     final int         surroundBefore;
     final int         surroundAfter;
+    final int         minCigarElementLength;
     final LevenshteinDistance levDistance = new LevenshteinDistance();
     final Integer     smqSize;
     final Integer     smqSizeMean;
-    final boolean     reportAllAlts;
-    final boolean     tagBasesWithAdjacentRefDiff;
-    final boolean     generateNonRefFeatures;
 
-    boolean     ignoreSurround;
-    int         spanBefore;
-    int         spanAfter;
-    int         minCigarElementLength;
+    final boolean     ignoreSurround;
+    final int         spanBefore;
+    final int         spanAfter;
+
+    final FlowFeatureMapperArgumentCollection fmArgs;
 
     public SNVMapper(FlowFeatureMapperArgumentCollection fmArgs) {
         surroundBefore = fmArgs.snvIdenticalBases;
         surroundAfter = (fmArgs.snvIdenticalBasesAfter != 0) ?  fmArgs.snvIdenticalBasesAfter : surroundBefore;
         smqSize = fmArgs.surroundingMediaQualitySize;
         smqSizeMean = fmArgs.surroundingMeanQualitySize;
-        reportAllAlts = fmArgs.reportAllAlts;
-        tagBasesWithAdjacentRefDiff = fmArgs.tagBasesWithAdjacentRefDiff;
-        generateNonRefFeatures = false;
-
-        init();
-    }
-
-    private void init() {
+        this.fmArgs = fmArgs;
 
         // ignore surround
-        ignoreSurround = reportAllAlts || tagBasesWithAdjacentRefDiff;
+        ignoreSurround = fmArgs.reportAllAlts || fmArgs.tagBasesWithAdjacentRefDiff;
         spanBefore = ignoreSurround ? 0 : surroundBefore;
         spanAfter = ignoreSurround ? 0 : surroundAfter;
         minCigarElementLength = spanBefore + 1 + spanAfter;
@@ -60,10 +52,10 @@ public class SNVMapper implements FeatureMapper {
     }
 
     @Override
-    public void forEachOnRead(GATKRead read, ReferenceContext referenceContext, Consumer<? super MappedFeature> action) {
+    public void forEachOnRead(GATKRead read, ReferenceContext referenceContext, Consumer<? super FlowFeatureMapper.MappedFeature> action) {
 
         // prepare list
-        List<MappedFeature>     features = new LinkedList<>();
+        List<FlowFeatureMapper.MappedFeature>     features = new LinkedList<>();
 
         // access bases
         final byte[]      bases = read.getBasesNoCopy();
@@ -78,15 +70,13 @@ public class SNVMapper implements FeatureMapper {
         } else {
             basesString = new String(Arrays.copyOfRange(bases, startSoftClip, bases.length - endSoftClip));
         }
-        int               refEditDistance = calcEditDistance(basesString, new String(ref));
+        int               refEditDistance = levDistance.apply(basesString, new String(ref));
 
         // count bases delta on M cigar elements
         int         nonIdentMBases = 0;
         int         readOfs = 0;
         int         refOfs = 0;
-        int         cigarElementCount = read.getCigar().numCigarElements();
-        for ( int cigarElementIndex = 0 ; cigarElementIndex < cigarElementCount ; cigarElementIndex++ ) {
-            final CigarElement cigarElement = read.getCigar().getCigarElement(cigarElementIndex);
+        for ( final CigarElement cigarElement : read.getCigarElements() ) {
             final int     length = cigarElement.getLength();
             if ( cigarElement.getOperator().consumesReadBases() && cigarElement.getOperator().consumesReferenceBases() ) {
                 for ( int ofs = 0 ; ofs < length ; ofs++ ) {
@@ -107,22 +97,19 @@ public class SNVMapper implements FeatureMapper {
         // walk the cigar (again) looking for features
         readOfs = 0;
         refOfs = 0;
-        int numCigarElements = read.numCigarElements();
-        for ( int cigarElementIndex = 0 ; cigarElementIndex < numCigarElements ; cigarElementIndex++ ) {
-            final CigarElement cigarElement = read.getCigarElement(cigarElementIndex);
+        for ( final CigarElement cigarElement : read.getCigarElements() ) {
+
             final int     length = cigarElement.getLength();
 
             // worth looking into?
             if ( length >= minCigarElementLength &&
                     cigarElement.getOperator().consumesReadBases() &&
-                    (generateNonRefFeatures || cigarElement.getOperator().consumesReferenceBases()) ) {
+                    cigarElement.getOperator().consumesReferenceBases() ) {
                 readOfs += spanBefore;
                 refOfs += spanBefore;
-                final boolean hasRef = cigarElement.getOperator().consumesReferenceBases();
-                for ( int ofs = spanBefore ; ofs < length - spanAfter ; ofs++, readOfs++, refOfs += (hasRef) ? 1 : 0 ) {
+                for ( int ofs = spanBefore ; ofs < length - spanAfter ; ofs++, readOfs++, refOfs++ ) {
 
-                    final byte refBase = hasRef ? ref[refOfs] : bases[readOfs];
-                    if ( refBase != 'N' && (reportAllAlts || (bases[readOfs] != refBase)) ) {
+                    if ( ref[refOfs] != 'N' && (fmArgs.reportAllAlts || (bases[readOfs] != ref[refOfs])) ) {
 
                         // check that this is really a SNV (must be surrounded by identical ref)
                         boolean     surrounded = true;
@@ -133,7 +120,7 @@ public class SNVMapper implements FeatureMapper {
                                 surrounded = false;
                                 continue;
                             }
-                            if ( !hasRef || (bases[bIndex] != ref[rIndex]) ) {
+                            if ( bases[bIndex] != ref[rIndex] ) {
                                 surrounded = false;
                             }
                         }
@@ -144,17 +131,17 @@ public class SNVMapper implements FeatureMapper {
                                 surrounded = false;
                                 continue;
                             }
-                            if ( !hasRef || (bases[bIndex] != ref[rIndex]) ) {
+                            if ( bases[bIndex] != ref[rIndex] ) {
                                 surrounded = false;
                             }
                         }
-                        if ( (!reportAllAlts && !tagBasesWithAdjacentRefDiff) && !surrounded ) {
+                        if ( (!fmArgs.reportAllAlts && !fmArgs.tagBasesWithAdjacentRefDiff) && !surrounded ) {
                             continue;
                         }
 
                         // add this feature
-                        MappedFeature feature = MappedFeature.makeSNV(read, readOfs, refBase, referenceContext.getStart() + refOfs, readOfs - refOfs);
-                        if ( (reportAllAlts || tagBasesWithAdjacentRefDiff) && !surrounded )
+                        FlowFeatureMapper.MappedFeature feature = FlowFeatureMapper.MappedFeature.makeSNV(read, readOfs, ref[refOfs], referenceContext.getStart() + refOfs, readOfs - refOfs);
+                        if ( (fmArgs.reportAllAlts || fmArgs.tagBasesWithAdjacentRefDiff) && !surrounded )
                             feature.adjacentRefDiff = true;
                         feature.nonIdentMBasesOnRead = nonIdentMBases;
                         feature.refEditDistance = refEditDistance;
@@ -201,6 +188,7 @@ public class SNVMapper implements FeatureMapper {
 
                         }
 
+
                         features.add(feature);
                     }
                 }
@@ -220,44 +208,10 @@ public class SNVMapper implements FeatureMapper {
         };
 
         // report features
-        for ( MappedFeature feature : features ) {
+        for ( FlowFeatureMapper.MappedFeature feature : features ) {
             feature.featuresOnRead = features.size();
             action.accept(feature);
         }
-    }
-
-    private int calcEditDistance(String s1, String s2) {
-
-        // identical
-        if ( s1.length() == s2.length() && s1.equals(s2) ) {
-            return 0;
-        }
-
-        // find first difference & trim
-        int     minLength = Math.min(s1.length(), s2.length());
-        int     diffIndex = 0;
-        for ( ; diffIndex < minLength ; diffIndex++ ) {
-            if (s1.charAt(diffIndex) != s2.charAt(diffIndex)) {
-                break;
-            }
-        }
-        s1 = s1.substring(diffIndex);
-        s2 = s2.substring(diffIndex);
-        minLength = Math.min(s1.length(), s2.length());
-
-        // find last difference & trim
-        int diffLength1 = s1.length();
-        int diffLength2 = s2.length();
-        for ( ; diffLength1 > 0 && diffLength2 > 0 ; diffLength1--, diffLength2-- ) {
-            if (s1.charAt(diffLength1 - 1) != s2.charAt(diffLength2 - 1)) {
-                break;
-            }
-        }
-        s1 = s1.substring(0, diffLength1);
-        s2 = s2.substring(0, diffLength2);
-
-        // find edit distance on the shorter strings
-        return levDistance.apply(s1, s2);
     }
 
     private int calcSmq(final byte[] quals, int from, int to, boolean median) {
@@ -303,9 +257,7 @@ public class SNVMapper implements FeatureMapper {
         // walk the cigar
         int         readOfs = 0;
         int         refOfs = 0;
-        int numCigarElements = read.numCigarElements();
-        for ( int cigarElementIndex = 0 ; cigarElementIndex < numCigarElements ; cigarElementIndex++ ) {
-            final CigarElement cigarElement = read.getCigarElement(cigarElementIndex);
+        for ( final CigarElement cigarElement : read.getCigarElements() ) {
 
             final int     length = cigarElement.getLength();
