+++ src/test/java/org/broadinstitute/hellbender/tools/walkers/groundtruth/GroundTruthScorerIntegrationTest.java
@@ -15,7 +15,6 @@ public class GroundTruthScorerIntegrationTest extends CommandLineProgramTest {
     public static final boolean UPDATE_EXACT_MATCH_EXPECTED_OUTPUTS = false;
 
     public static final String OUTPUT_FILENAME = "ground_truth_scorer_output.csv";
-    public static final String OUTPUT_FILENAME_SET_MIN_ERROR_PROB = "ground_truth_scorer_output_preset_filling_prob.csv";
     public static final String OUTPUT_FILENAME_MEANCALL = "ground_truth_scorer_output_meancall.csv";
     public static final String OUTPUT_FILENAME_UNCLIPPED = "ground_truth_scorer_output_unclipped.csv";
     public static final String OUTPUT_FILENAME_GENOME_PRIOR = "ground_truth_scorer_output_genome_prior.csv";
@@ -52,27 +51,7 @@ public class GroundTruthScorerIntegrationTest extends CommandLineProgramTest {
             IntegrationTestSpec.assertEqualTextFiles(outputFile, expectedFile);
         }
     }
-    @Test
-    public void testBasicWithExternalFillingValue() throws IOException {
 
-        final File outputDir = createTempDir("testGroundTruthTest");
-        final File expectedFile = new File(testDir + "/" + OUTPUT_FILENAME_SET_MIN_ERROR_PROB);
-        final File outputFile = UPDATE_EXACT_MATCH_EXPECTED_OUTPUTS ? expectedFile : new File(outputDir + "/" + OUTPUT_FILENAME);
-
-        final String[] args = ArrayUtils.addAll(buildCommonArgs(outputFile, GT_SCORER_INPUT_BAM, true),
-                "--"+GroundTruthScorer.SET_MIN_ERROR_PROB);
-
-
-        runCommandLine(args);  // no assert, just make sure we don't throw
-
-        // make sure we've generated the output file
-        Assert.assertTrue(outputFile.exists());
-
-        // walk the output and expected files, compare non-comment lines
-        if ( !UPDATE_EXACT_MATCH_EXPECTED_OUTPUTS ) {
-            IntegrationTestSpec.assertEqualTextFiles(outputFile, expectedFile);
-        }
-    }
     @Test
     public void testMeanCall() throws IOException {
 
@@ -197,7 +176,7 @@ public class GroundTruthScorerIntegrationTest extends CommandLineProgramTest {
                 "-I", testDir + "/" + inputFile,
                 "--output-csv", outputFile.getAbsolutePath(),
                 "--intervals", smallInterval ? CHR9_SMALL_INTERVAL : CHR9_ALL_INTERVAL,
-                "--likelihood-calculation-engine", "FlowBased",
+                "--likelihood-calculation-engine", "FlowBased"
         };
 
     }
