--- src/test/resources/large/featureMapping/snv_feature_mapper_AddApplySNVQRFeatures.vcf
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:f3349c6e71f4f4661eef88b27cabd29e652a0d260ac6a730fec0398c02cbcd4e
-size 348895
