--- src/test/resources/large/featureMapping/snv_feature_mapper_AddApplySNVQRFeatures.vcf.idx
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:a0dc0f3dcd2370191d746f310ff380a64dfb278f5d584402c8d0f8ec59b35dd7
-size 118526
