+++ src/test/resources/large/featureMapping/snv_feature_mapper_output.vcf
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:bf2e8495a19e7711eeabbf2c2bf02d337d608159f9a05f8d1c0bf76f3d56407d
-size 6223385
+oid sha256:b5f7a63955d57c04eac81f19140ea3483f93943e0a4fdb0990e636f3c08fbda1
+size 5722270
