+++ src/test/resources/large/featureMapping/snv_feature_mapper_output.vcf.idx
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:c5c444aa9e4445e1ea7b3b67a2a26f9e321e53c1f5b36feefec399b5ce16ba86
+oid sha256:35aac2c14df674d928a33d0faa938baecb3f67aa60d9a49f4f6f0f290683dc4c
 size 120113
