+++ src/test/resources/large/featureMapping/snv_feature_mapper_qc_failed_output.vcf
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:b9aaa7aba507cb6326a239e631c6d131af626c5b01cadd18e1d1f1f4b8a8249d
-size 5553047
+oid sha256:405fb1bf62168e22c504e0e4ae0d9d6b9ab4f7ac32ff18dcd77f071618d964da
+size 5051930
