+++ src/test/resources/large/featureMapping/snv_feature_mapper_qc_failed_output.vcf.idx
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:b7dcdd60aa547f430034853716ef676c30919f732285cad34a0d15d1412391c2
+oid sha256:7aba6151d662570d8e4a25cbb4717c65009c2d217375a287667b1f47c2db9925
 size 120123
