+++ src/test/resources/large/featureMapping/snv_feature_mapper_report_all_alts_output.vcf
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:08b86c91288c58074670a0be9d6142022c1ba2d47971e283da52334429f1b8cd
-size 23193116
+oid sha256:1196abbe9d96b3d48dfae5a37991aaacaf2474a4bcae5c36cfae58378f8c1429
+size 21938650
