+++ src/test/resources/large/featureMapping/snv_feature_mapper_report_all_alts_output.vcf.idx
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:bc017062d491fa168a210b094ad9b1d33df689b307aba593947cebbff99fcf2f
+oid sha256:93e8a05eece74ee9760982366b0234cc2d4ee4ac25eb7e26f89b0a106958e235
 size 134797
