+++ src/test/resources/large/featureMapping/snv_feature_mapper_smq_all_output.vcf
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:816a00f48f32753cad87e3b34edc8630c7c085f47fb6704318ad23bd76a7861a
-size 7256207
+oid sha256:4245184ac90c7dc1a0968285ddb0878c2a9d46f27b8cc1dd71b11074fd477d6e
+size 6755057
