+++ src/test/resources/large/featureMapping/snv_feature_mapper_smq_all_output.vcf.idx
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:94a9767aff001e998e96e6cf2e07651d27c0113743d094c566894e4cbb36c90c
+oid sha256:bdad1cfb039ec6d61f74a739d889adb65a1920a7a06c1e114522e7f229aacf84
 size 120121
