+++ src/test/resources/large/featureMapping/snv_feature_mapper_smq_output.vcf
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:e20cffa3f1806f3964209b16ed5a061b8014d5721eb9bb0e6299e20f3e32e61a
-size 6279385
+oid sha256:694325ff66b4f1184593f22b5fc3171f6609e55286763c49b66fe7bc2f68275f
+size 5778235
