+++ src/test/resources/large/featureMapping/snv_feature_mapper_smq_output.vcf.idx
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:9089f823744e3ff214269411cccb54c6e50640329f2e6722a6f8366889e0efea
+oid sha256:0a3dac7c3980303af8f127a4cfcc98343f7e9a8be4755327237f4cdc155c99f2
 size 120117
