+++ src/test/resources/large/featureMapping/snv_feature_mapper_tag_adjs_output.vcf
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:bbcef0ad75fc81887a6a9057fd204e06f6619b1ff6ef73c242717f4f24782d75
-size 251630
+oid sha256:61ce347a59b8a814c5af0bc08f242590feffbd7fc010cbc948cba94f35d6437d
+size 250644
