+++ src/test/resources/large/featureMapping/snv_feature_mapper_tag_adjs_output.vcf.idx
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:b74d64093ca56a3cd58bc7fc478fb7e34ba2033a4a900feff96c858791d26679
+oid sha256:c9876e614b5e5d9e3ffb1d85207149210e931a241821c832b3f3efc095b59710
 size 118520
