+++ src/test/resources/large/groundTruth/ground_truth_scorer_output.csv
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:d58e2278ce68f744d42c73966489bee78183c9f9ff43f8fb1a230e1f251a2867
-size 129401
+oid sha256:4d8963b8543fdd8522da78ed09560a67f3d6658638a49ed59d4c587c85738cf6
+size 121169
