+++ src/test/resources/large/groundTruth/ground_truth_scorer_output_filter.csv
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:5586c78009680511d0dc7d3663d71e9e0791e63837df66c8767cbd4233758c5c
-size 125404
+oid sha256:140e38e827b31101b0f1e881ad7f14524110d51f6f5189f7856febb0a5e3e17e
+size 117442
