+++ src/test/resources/large/groundTruth/ground_truth_scorer_output_genome_prior.csv
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:9200b6afeafd44c34a292c11cfbb8fc1a6c115df4c2870000860ce2596f1c13e
-size 137483
+oid sha256:e28a63a0512375acfeaed02080e0fd12ae2f55e9ec1b473e7a6e5741bd5df5c5
+size 145408
