+++ src/test/resources/large/groundTruth/ground_truth_scorer_output_meancall.csv
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:6549290df6e769e7c2ac158678ba720011b37c05ae9ef83c013b568aa18676b2
-size 1247645
+oid sha256:2d9cdf557ff28002661d4d68a26950c2802d4f1f6b7580ed30e2a494b8c9a9f0
+size 897467
