--- src/test/resources/large/groundTruth/ground_truth_scorer_output_preset_filling_prob.csv
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:4d8963b8543fdd8522da78ed09560a67f3d6658638a49ed59d4c587c85738cf6
-size 121169
