+++ src/test/resources/large/groundTruth/ground_truth_scorer_output_report.txt
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:014f820c97103a78dd61aa67bcaf05518172102f865169af1a191b08984f05b7
-size 73138
+oid sha256:d3e276968d778867340599d20be701a08ba14440614f7be8cbe9add7569a6c4e
+size 65675
