+++ src/test/resources/large/groundTruth/ground_truth_scorer_output_unclipped.csv
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:9f551ba6887e2dd0ed979160f079018e041114d7141e9ec004976f495906d9a1
-size 89837
+oid sha256:549f8c2eb02aa1aad1052e2653a5d009134592c7fdec9ace9d21957d3941a40c
+size 95335
